+++ src/Resources/config/routing/all.xml
@@ -58,8 +58,7 @@
             prefix="/wysiwyg" />
     <import resource="@EMSCoreBundle/Resources/config/routing/wysiwyg-options.xml"
             prefix="/wysiwyg-options" />
-    <import resource="@EMSCoreBundle/Resources/config/routing/user.xml"
-            prefix="/user" />
+    <import resource="@EMSCoreBundle/Resources/config/routing/user/user.xml" />
     <import resource="@EMSCoreBundle/Resources/config/routing/profile.xml"
             prefix="/profile" />
     <import resource="@EMSCoreBundle/Resources/config/routing/log.xml"
