--- src/Resources/config/routing/user.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<routes xmlns="http://symfony.com/schema/routing"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://symfony.com/schema/routing http://symfony.com/schema/routing/routing-1.0.xsd">
-
-    <route id="emsco_user_datatable" path="/datatable.json" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:ajaxDataTableAction</default>
-    </route>
-    <route id="emsco_user_index" path="/" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:indexAction</default>
-    </route>
-    <route id="emsco_user_add" path="/add" methods="GET POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:addUserAction</default>
-    </route>
-    <route id="emsco_user_edit" path="/edit/{user}" methods="GET POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:edit</default>
-    </route>
-    <route id="emsco_user_delete" path="/delete/{user}" methods="POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:delete</default>
-    </route>
-    <route id="emsco_user_enabling" path="/enabling/{user}" methods="POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:enabling</default>
-    </route>
-    <route id="emsco_user_api_key" path="/api-key/{username}" methods="POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:apiKey</default>
-    </route>
-    <route id="emsco_user_spreadsheet_export" path="/users.{_format}" methods="GET">
-        <requirement key="_format">csv|xlsx</requirement>
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:spreadsheetExport</default>
-    </route>
-
-    <route id="emsco_user_profile" path="/profile" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\User\ProfileController::show</default>
-    </route>
-    <route id="emsco_user_profile_edit" path="/profile/edit" methods="GET POST">
-        <default key="_controller">EMS\CoreBundle\Controller\User\ProfileController::edit</default>
-    </route>
-    <route id="emsco_user_change_password" path="/profile/change-password" methods="GET POST">
-        <default key="_controller">EMS\CoreBundle\Controller\User\ChangePasswordController::changePassword</default>
-    </route>
-
-
-    <!-- Deprecated -->
-    <route id="ems_core_user_ajax_data_table" path="/datatable.json" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:ajaxDataTableAction</default>
-    </route>
-    <route id="ems.user.index" path="/" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:indexAction</default>
-    </route>
-    <route id="user.add" path="/add" methods="GET POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:addUserAction</default>
-    </route>
-    <route id="user.edit" path="/{id}/edit" methods="GET POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:editUserAction</default>
-    </route>
-    <route id="user.delete" path="/{id}/delete" methods="POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:removeUserAction</default>
-    </route>
-    <route id="user.enabling" path="/{id}/enabling" methods="POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:enablingUserAction</default>
-    </route>
-    <route id="EMS_user_apikey" path="/{username}/apikey" methods="POST">
-        <default key="_controller">EMS\CoreBundle\Controller\UserController:apiKeyAction</default>
-    </route>
-
-</routes>
