+++ src/Resources/config/routing/user/login.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<routes xmlns="http://symfony.com/schema/routing"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://symfony.com/schema/routing http://symfony.com/schema/routing/routing-1.0.xsd">
+
+    <route id="fos_user_security_login" path="/login" methods="GET POST">
+        <default key="_controller">EMS\CoreBundle\Controller\User\LoginController::login</default>
+    </route>
+    <route id="fos_user_security_check" path="/login_check" methods="POST">
+        <default key="_controller">EMS\CoreBundle\Controller\User\LoginController::check</default>
+    </route>
+    <route id="fos_user_security_logout" path="/logout" methods="GET POST">
+        <default key="_controller">EMS\CoreBundle\Controller\User\LoginController::logout</default>
+    </route>
+
+</routes>
