+++ src/Resources/config/routing/user/resetting.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<routes xmlns="http://symfony.com/schema/routing"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://symfony.com/schema/routing http://symfony.com/schema/routing/routing-1.0.xsd">
+
+    <route id="emsco_user_resetting_request" path="/resetting/request" methods="GET POST">
+        <default key="_controller">EMS\CoreBundle\Controller\User\ResettingController::request</default>
+    </route>
+    <route id="emsco_user_resetting_check_email" path="/resetting/check_email" methods="GET">
+        <default key="_controller">EMS\CoreBundle\Controller\User\ResettingController::checkEmail</default>
+    </route>
+    <route id="emsco_user_resetting_reset" path="/resetting/reset/{token}" methods="GET POST">
+        <default key="_controller">EMS\CoreBundle\Controller\User\ResettingController::reset</default>
+    </route>
+
+</routes>
